#6701 Report storage tests (#6814)
* #6701 Report storage tests

Functional and unit tests

* #6701 Test cleanup

* #6701 js-storage teardown

The first time js-storage is required it evaluates if it's running in the browser or not: https://github.com/julien-maurel/js-storage/blob/master/js.storage.js#L423 and will define the localstorage getters and setters accordingly. This becomes an issue if testing localstorage between UI and non-UI tests. reportstorage.test.js was requiring it before hashauth.test.js causing a conflict.

* #6701 false positive test

The page isn't refreshing

Co-authored-by: Sulka Haro <sulka@sulka.net>